refactor(rust): split files past the 200-line soft cap
