+++ src-tauri/src/sync/git/backend.rs
@@ -56,7 +56,7 @@ impl GitBackend {
     /// push. Delegates the cred-shape selection to the free function
     /// returned by [`make_credentials_callback`] so the closure body
     /// itself can be exercised without a real authenticated remote.
-    fn credentials_callbacks(&self) -> RemoteCallbacks<'_> {
+    pub(super) fn credentials_callbacks(&self) -> RemoteCallbacks<'_> {
         let mut cb = RemoteCallbacks::new();
         cb.credentials(make_credentials_callback(self.https_token.clone()));
         cb
@@ -77,205 +77,6 @@ impl GitBackend {
             .get_string("user.email")
             .unwrap_or_else(|_| format!("{}@localhost", crate::APP_NAME));
         git2::Signature::now(&name, &email).map_err(|e| SyncError::Backend(e.message().to_string()))
-    }
-
-    /// True if the working tree differs from HEAD in any visible way.
-    fn working_tree_dirty(repo: &Repository) -> Result<bool, SyncError> {
-        let mut opts = StatusOptions::new();
-        opts.include_untracked(true);
-        opts.recurse_untracked_dirs(true);
-        let statuses = repo
-            .statuses(Some(&mut opts))
-            .map_err(|e| SyncError::Backend(e.message().to_string()))?;
-        Ok(!statuses.is_empty())
-    }
-
-    /// Stage every change in the working tree (mirror of `git add -A`).
-    pub(super) fn stage_all(repo: &Repository) -> Result<bool, SyncError> {
-        let mut index = repo
-            .index()
-            .map_err(|e| SyncError::Backend(e.message().to_string()))?;
-        index
-            .add_all(["*"].iter(), git2::IndexAddOption::DEFAULT, None)
-            .map_err(|e| SyncError::Backend(e.message().to_string()))?;
-        index
-            .write()
-            .map_err(|e| SyncError::Backend(e.message().to_string()))?;
-
-        // Did the index actually move? Returns true when there is
-        // something staged that would produce a real commit.
-        let head_tree = match repo.head() {
-            Ok(head) => head
-                .peel_to_tree()
-                .map_err(|e| SyncError::Backend(e.message().to_string()))?,
-            Err(e) if e.code() == git2::ErrorCode::UnbornBranch => {
-                // No commits yet — anything in the index counts as a
-                // change.
-                return Ok(!index.is_empty());
-            }
-            // Defensive: libgit2's `repo.head()` either succeeds, returns
-            // `UnbornBranch` (handled above), or surfaces a lower-level
-            // I/O failure (corrupt refs file etc.) we don't try to
-            // recover from. Not reachable in normal use, so no test.
-            Err(e) => return Err(SyncError::Backend(e.message().to_string())),
-        };
-        let index_tree_oid = index
-            .write_tree()
-            .map_err(|e| SyncError::Backend(e.message().to_string()))?;
-        let index_tree = repo
-            .find_tree(index_tree_oid)
-            .map_err(|e| SyncError::Backend(e.message().to_string()))?;
-        let diff = repo
-            .diff_tree_to_tree(Some(&head_tree), Some(&index_tree), None)
-            .map_err(|e| SyncError::Backend(e.message().to_string()))?;
-        Ok(diff.deltas().len() > 0)
-    }
-
-    fn commit_index(
-        repo: &Repository,
-        signature: &git2::Signature<'_>,
-        message: &str,
-        parents: Vec<git2::Commit<'_>>,
-    ) -> Result<git2::Oid, SyncError> {
-        let mut index = repo
-            .index()
-            .map_err(|e| SyncError::Backend(e.message().to_string()))?;
-        let tree_oid = index
-            .write_tree()
-            .map_err(|e| SyncError::Backend(e.message().to_string()))?;
-        let tree = repo
-            .find_tree(tree_oid)
-            .map_err(|e| SyncError::Backend(e.message().to_string()))?;
-        let parent_refs: Vec<&git2::Commit<'_>> = parents.iter().collect();
-        repo.commit(
-            Some("HEAD"),
-            signature,
-            signature,
-            message,
-            &tree,
-            &parent_refs,
-        )
-        .map_err(|e| SyncError::Backend(e.message().to_string()))
-    }
-
-    fn head_commit(repo: &Repository) -> Result<Option<git2::Commit<'_>>, SyncError> {
-        match repo.head() {
-            Ok(head) => head
-                .peel_to_commit()
-                .map(Some)
-                .map_err(|e| SyncError::Backend(e.message().to_string())),
-            Err(e) if e.code() == git2::ErrorCode::UnbornBranch => Ok(None),
-            // Same defensive fallthrough as `stage_all` — non-`UnbornBranch`
-            // `repo.head()` failures are corrupt-repo territory; we surface
-            // a Backend error and let the user re-clone.
-            Err(e) => Err(SyncError::Backend(e.message().to_string())),
-        }
-    }
-
-    /// Return (ahead, behind) of the local branch relative to its
-    /// remote-tracking branch. Returns (0, 0) when there is no upstream
-    /// configured yet (fresh init before the first push).
-    fn ahead_behind(repo: &Repository, branch_name: &str) -> Result<(u32, u32), SyncError> {
-        let local = match repo.find_branch(branch_name, BranchType::Local) {
-            Ok(b) => b,
-            Err(_) => return Ok((0, 0)),
-        };
-        let upstream_ref = format!("refs/remotes/{ORIGIN}/{branch_name}");
-        let upstream = match repo.find_reference(&upstream_ref) {
-            Ok(r) => r,
-            Err(_) => return Ok((0, 0)),
-        };
-        let local_oid = local
-            .get()
-            .target()
-            .ok_or_else(|| SyncError::InvalidState("local branch has no tip".into()))?;
-        let upstream_oid = upstream
-            .target()
-            .ok_or_else(|| SyncError::InvalidState("upstream ref has no tip".into()))?;
-        repo.graph_ahead_behind(local_oid, upstream_oid)
-            .map(|(a, b)| (a as u32, b as u32))
-            .map_err(|e| SyncError::Backend(e.message().to_string()))
-    }
-
-    fn collect_conflicts(repo: &Repository) -> Result<Vec<String>, SyncError> {
-        let mut out = Vec::new();
-        let index = repo
-            .index()
-            .map_err(|e| SyncError::Backend(e.message().to_string()))?;
-        if !index.has_conflicts() {
-            return Ok(out);
-        }
-        let conflicts = index
-            .conflicts()
-            .map_err(|e| SyncError::Backend(e.message().to_string()))?;
-        for entry in conflicts {
-            let entry = entry.map_err(|e| SyncError::Backend(e.message().to_string()))?;
-            // Prefer the "ours" path; fall back to "theirs"/"ancestor".
-            let path = entry
-                .our
-                .as_ref()
-                .or(entry.their.as_ref())
-                .or(entry.ancestor.as_ref())
-                .map(|e| String::from_utf8_lossy(&e.path).into_owned());
-            if let Some(p) = path {
-                out.push(p);
-            }
-        }
-        out.sort();
-        out.dedup();
-        Ok(out)
-    }
-
-    /// Whether a remote is configured for this workspace. An empty
-    /// `remote_url` means local-only sync: Glyph still stages and commits
-    /// changes into the repo's history, but there is nothing to fetch,
-    /// merge, or push.
-    fn has_remote(&self) -> bool {
-        !self.config.remote_url.trim().is_empty()
-    }
-
-    fn fetch_remote(&self, repo: &Repository) -> Result<(), SyncError> {
-        let mut remote = repo
-            .find_remote(ORIGIN)
-            .map_err(|_| SyncError::NotConfigured)?;
-        let mut opts = FetchOptions::new();
-        opts.remote_callbacks(self.credentials_callbacks());
-        let refspec = format!(
-            "refs/heads/{0}:refs/remotes/{ORIGIN}/{0}",
-            self.config.remote_branch
-        );
-        remote
-            .fetch(&[&refspec], Some(&mut opts), None)
-            .map_err(map_remote_error)?;
-        Ok(())
-    }
-
-    fn push_branch(&self, repo: &Repository) -> Result<(), SyncError> {
-        let mut remote = repo
-            .find_remote(ORIGIN)
-            .map_err(|_| SyncError::NotConfigured)?;
-        let mut opts = PushOptions::new();
-        opts.remote_callbacks(self.credentials_callbacks());
-        let refspec = format!("refs/heads/{0}:refs/heads/{0}", self.config.remote_branch);
-        remote
-            .push(&[&refspec], Some(&mut opts))
-            .map_err(map_remote_error)?;
-        Ok(())
-    }
-
-    /// Whether the `.glyph/` config directory is already present in HEAD.
-    /// `false` on an unborn branch (no commits yet).
-    fn config_tracked(repo: &Repository) -> Result<bool, SyncError> {
-        let head_tree = match repo.head() {
-            Ok(h) => h
-                .peel_to_tree()
-                .map_err(|e| SyncError::Backend(e.message().to_string()))?,
-            Err(e) if e.code() == git2::ErrorCode::UnbornBranch => return Ok(false),
-            Err(e) => return Err(SyncError::Backend(e.message().to_string())),
-        };
-        Ok(head_tree
-            .get_path(Path::new(&format!(".{}", crate::APP_NAME)))
-            .is_ok())
     }
 }
